refactor: organize and consolidate type definitions
- Create centralized types file for shared types and interfaces
- Move component-specific props back to their respective component files
- Keep shared types (callbacks, state, actions) in centralized location
- Improve maintainability while respecting component encapsulation
