+++ frontend/src/features/skillChaining/skills/hooks/__tests__/useSkillGraph.test.tsx
@@ -1,4 +1,4 @@
-import { describe, it, expect, vi } from 'vitest';
+import { describe, it, expect, vi, beforeEach } from 'vitest';
 import { renderHook, waitFor } from '@testing-library/react';
 import { MockedProvider } from '@apollo/client/testing';
 import { GET_LINKED_SKILLS } from '@api/graphql/queries';
